one of the commits of all time

+++ Train_Controller_SW/trainControllerSW.py
@@ -1,5 +1,6 @@
 import linecache
 from SystemTime import SystemTime
+import time
 from PyQt6.QtCore import pyqtSignal
 
 
@@ -26,7 +27,7 @@ class TrainController:
 
     # constants
     maxSpeed = 19.44  # 70 kmh in m/s
-    maxPower = 120000  # in watts
+    maxPower = float(120000)  # in watts
     accelLim = float(0.5)  # acceleration limit of the train based on 2/3 load
     decelLim = float(-1.2)  # deceleration limit of the train based on 2/3 load
     eBrakeDecelLim = float(-2.73)  # deceleration limit of the train emergency brake based on current capacity
@@ -50,6 +51,8 @@ class TrainController:
         self.doorState = bool(0)  # door open/close status as bool: 0 closed, 1 opened (default 0)
         self.doorL = bool(0)  # door control for left side doors: 0 closed, 1 opened (default 0)
         self.doorR = bool(0)  # door control for right side doors: 0 close, 1 opened (default 0)
+        self.extLightsByDriver = bool(False)
+        self.intLightsByDriver = bool(False)
 
         # cabin environment, as inputs from train model
         self.cabinTemp = float(68)  # cabin thermostat temp as float, in Fahrenheit (default 68)
@@ -67,7 +70,7 @@ class TrainController:
         # train model inputs
         self.actualSpeed = float(0)  # actual speed of the train
         self.cmdSpeed = float(0)  # commanded vital speed of the train, as passed from wayside controller
-        # self.speedLim = float(31.0686)  # speed limit of the current block, 50 kmh in mph, blue line speed lim
+        self.speedLim = float(31.0686)  # speed limit of the current block, 50 kmh in mph, blue line speed lim
         self.vitalAuth = float(0)  # vital authority for the train, as passed from wayside controller
         self.passEBrake = bool(0)  # passenger emergency brake as bool: 0 off, 1 on
         self.polarity = bool(0)  # track circuit state: 0 left, 1 right
@@ -77,10 +80,10 @@ class TrainController:
         # self.currSpeed = float(0.5)  # referenced current speed from actual speed of train model
         self.setPtSpeed = float(0)  # set point speed as determined by driver input
         self.power = float(0)  # power of the train engine as determined by driver and engineer input
-        self.servBrake = bool(0)
-        self.eBrake = bool(0)  # application of emergency brake as bool: 0 off, 1 on
+        self.servBrake = bool(False)
+        self.eBrake = bool(False)  # application of emergency brake as bool: 0 off, 1 on
 
-        self.prevSpeed = float(0)  # velocity of previous step
+        self.sBrakeSetByDriver = bool(False)
 
         # power calculation vars
         self.uk = float(0)
@@ -89,8 +92,8 @@ class TrainController:
         self.ek1 = float(0)
         self.T = 0  # time between sample from train model(eg time between train controller value updates for IT3)
         # PID gain values, as inputs from engineer
-        self.ki = float(50)  # integral gain as float,
-        self.kp = float(50)  # proportional gain as float
+        self.ki = float(500)  # integral gain as float,
+        self.kp = float(1000)  # proportional gain as float
 
         # authority and distance
         self.distanceTraveled = float(-25)
@@ -104,7 +107,8 @@ class TrainController:
         self.trainStopped = bool(1)
 
         self.makeAnnouncement = bool(0)  # if announcement is to be made: 0 no, 1 yes
-        self.station = bool(0)
+        self.atStation = bool(0)
+        self.station = ""
 
         self.i = int(1)  # index for reading through block list
         self.blockline = linecache.getline('Resources/IT3_GreenLine.txt', self.i).split("\t")
@@ -146,16 +150,19 @@ class TrainController:
         self.beacon = newbeacon
         return
 
-    def setcurrspeed(self, newcurrspeed):
-        self.currSpeed = newcurrspeed
+    def setactspeed(self, newactspeed):
+        if newactspeed >= self.speedlim:
+            self.actualSpeed = self.speedlim
+        else:
+            self.actualSpeed = newactspeed
         return
 
     def setsetptspeed(self, newsetptspeed):
         if newsetptspeed > self.cmdSpeed:  # new setpt speed cannot exceed command speed, cap at cmd speed
             newsetptspeed = self.cmdSpeed
 
-        if newsetptspeed > self.speedLim:  # new setpt speed cannot exceed speed lim, cap at speed lim
-            newsetptspeed = self.speedLim
+        if newsetptspeed > self.speedlim:  # new setpt speed cannot exceed speed lim, cap at speed lim
+            newsetptspeed = self.speedlim
 
         if newsetptspeed != self.setPtSpeed:  # TODO need to update engine power
             self.setPtSpeed = newsetptspeed
@@ -177,6 +184,9 @@ class TrainController:
             self.eBrake = 0
         elif self.eBrake == 0:
             self.eBrake = 1  # ebrake turns on
+            self.servBrake = False
+            self.sBrakeSetByDriver = False
+            print(self.servBrake)
             self.power = 0  # power cut
 
     def passebrakecontrol(self):
@@ -188,29 +198,35 @@ class TrainController:
         return
 
     def intlightscontrol(self):
-        if self.intLights == 0:
-            self.intLights = 1  # turn on cabin lights
-        elif self.intLights == 1:
-            self.intLights = 0  # turn off cabin lights
+        if not self.intLights:
+            self.intLights = True  # turn on cabin lights
+            self.intLightsByDriver = True
+        elif self.intLights:
+            self.intLights = False  # turn off cabin lights
+            self.intLightsByDriver = False
         return
 
     def intlightson(self):
-        self.intLights = 1
+        self.intLights = True
 
     def intlightsoff(self):
-        self.intLights = 0
+        self.intLights = False
 
     def extlightson(self):
-        self.extLights = 1
+        self.extLights = True
 
     def extlightsoff(self):
-        self.extLights = 0
+        self.extLights = False
 
     def extlightscontrol(self):
-        if self.extLights == 0:
-            self.extLights = 1  # turn on headlights
-        elif self.extLights == 1:
-            self.extLights = 0  # turn off headlights
+        if not self.extLights:
+            print("turned on")
+            self.extLights = True  # turn on headlights
+            self.extLightsByDriver = True
+        elif self.extLights:
+            print("turned off")
+            self.extLights = False  # turn off headlights
+            self.extLightsByDriver = False
         return
 
     def tempcontrol(self, temp):
@@ -256,12 +272,13 @@ class TrainController:
             self.section = self.blockline[1]
             self.block = self.blockline[2]
             self.length = float(self.blockline[3])
-            self.speedlim = float(self.blockline[4])
+            self.speedlim = float(self.blockline[4])/self.KMH2ms
             self.notes = self.blockline[5]
             door = self.blockline[6]
 
             if self.i == 172:
                 print("train controller sw.py: looped to yard")
+                print()
                 self.i = 0
 
             self.nextblockline = linecache.getline('Resources/IT3_GreenLine.txt', self.i+1).split("\t")
@@ -269,13 +286,16 @@ class TrainController:
             self.nextsection = self.blockline[1]
             self.nextblock = self.blockline[2]
             self.nextlength = float(self.blockline[3])
-            self.nextspeedlim = float(self.blockline[4])
+            self.nextspeedlim = float(self.blockline[4])/self.KMH2ms
             self.nextnotes = self.blockline[5]
 
             print(f'train controller sw.py: current block is: {self.block}')
-            if self.notes != "":
-                print(self.notes)
-                print(door)
+            if self.notes != "" and self.notes != "SPAWN":
+                # print(self.notes)
+                # print(door)
+                self.atStation = True
+                noteslist = self.notes.split("; ", -1)
+                self.station = noteslist[1]
                 if door == "Right\n" or door == "Right" or str(door) == "Right":
                     self.doorSide = 2
                 elif door == "Left\n":
@@ -284,11 +304,13 @@ class TrainController:
                     self.doorSide = 3
                 else:
                     self.doorSide = 0
+                print(f'station is: {self.station}')
+                print(f'train controller sw.py: door side is {self.doorSide}')
             else:
                 self.doorSide = 0
+                self.atStation = False
 
-            print(f'door side is {self.doorSide}')
-
+            self.arriveatstation(self.atStation)
         return
 
     def signalfailcontrol(self):
@@ -332,14 +354,15 @@ class TrainController:
 
         print(f'stopSoon: {self.stopSoon}')
         print(f'ebrake: {self.eBrake}')
+        print(f'service brake: {self.servBrake}')
         print(f'train safe to move: {self.trainSafeToMove}')
         # do power checks and calcs, train doesn't need to stop, brake not on, train can move
         # stopSoon: false, eBrake: false, trainSafeToMove: true
-        if not self.stopSoon and not self.eBrake and self.trainSafeToMove:
+        if not self.stopSoon and not self.eBrake and self.trainSafeToMove and not self.servBrake:
 
             if self.nextspeedlim < self.speedlim:
                 self.power = 0  # kill power
-                self.servBrake = 1  # enable service brake and slow down
+                self.servBrake = True  # enable service brake and slow down
             elif (self.nextspeedlim == self.speedlim) and (self.actualSpeed >= self.nextspeedlim):
                 self.power = 0  # kill power and coast
             else:
@@ -354,11 +377,23 @@ class TrainController:
                 self.uk = self.uk1 + (t/2)*(self.ek - self.ek1)
 
                 # power calculation thank you Dr. Profeta (Final Project vF.pdf, slides 61-66) (pages 54-59) :3
-                self.power = self.kp * self.ek + self.ki * self.uk
+                # self.power = self.kp * self.ek + self.ki * self.uk
+
+                # using TMR
+                power1 = self.kp * self.ek + self.ki * self.uk
+                power2 = self.kp * self.ek + self.ki * self.uk
+                power3 = self.kp * self.ek + self.ki * self.uk
+                self.power = power1 and power2 and power3
 
                 # check to make sure max power is not exceeded
                 if self.power > self.maxPower:
+                    print("capping at max")
                     self.power = self.maxPower
+
+                if self.power < 0:
+                    self.power = 0  # power can't be negative, enable service brake instead
+                    self.servBrake = True
+                    print("service brake on")
         else:
             self.power = 0
 
@@ -374,15 +409,22 @@ class TrainController:
 
     def authority(self):
         print("train controller sw.py: in authority func")
-        if self.vitalAuth < 4:
-            print(f'train controller sw.py: train authority less then 4, currently at {self.vitalAuth}')
-            self.stopSoon = True
-            self.power = 0
-            self.servBrake = True
-        elif self.vitalAuth >= 4:
-            self.stopSoon = False
-            self.servBrake = False
-            print("train controller sw.py: train authority has been given authority 4, able to move forward")
+        if self.vitalAuth is not None:
+            if self.vitalAuth < 4:
+                print(f'train controller sw.py: train authority less then 4, currently at {self.vitalAuth}')
+                self.stopSoon = True
+                self.power = 0
+                if not self.sBrakeSetByDriver:
+                    self.servBrake = True
+            elif self.vitalAuth >= 4:
+                self.stopSoon = False
+                if not self.sBrakeSetByDriver:
+                    self.servBrake = False
+                print("train controller sw.py: train authority has been given authority 4, able to move forward")
+        else:
+            self.vitalAuth = None
+            self.eBrake = True
+
 
         print("train controller sw.py: authority updated")
 
@@ -400,16 +442,23 @@ class TrainController:
         elif self.vitalAuth != 0 and self.cmdSpeed != 0:
             self.trainSafeToMove = True
             print("train controller sw.py: rain safe to move: authority and cmd speed received")
+        elif self.vitalAuth is None or self.cmdSpeed is None:
+            self.trainSafeToMove = False
 
         print()
 
     def speedcheck(self):
         # check if cmd speed and actual speed are within legal speed limits:
-        if self.cmdSpeed > self.speedlim:
-            print("train controller sw.py: command speed too high, limiting to speed limit")
-            self.cmdSpeed = self.speedlim
-        elif self.cmdSpeed <= self.speedlim:
-            print("train controller sw.py: command speed within speed limits")
+        if self.cmdSpeed is not None:
+            if self.cmdSpeed > self.speedlim:
+                print("train controller sw.py: command speed too high, limiting to speed limit")
+                self.cmdSpeed = self.speedlim
+            elif self.cmdSpeed <= self.speedlim:
+                print("train controller sw.py: command speed within speed limits")
+        else:
+            self.cmdSpeed = None
+            print("No speed picked up, signal pick up failure")
+            self.eBrake = True
 
         print()
 
@@ -426,8 +475,10 @@ class TrainController:
     def updater(self, inputs, num):
         print("train controller sw.py: train controller values being updated in back")
         print(inputs)
-        self.servBrake = 0  # turn off service brake, will be turned on if needed again
-        #print("serve brake turned off")
+
+        if not self.sBrakeSetByDriver:
+            self.servBrake = False  # turn off service brake, will be turned on if needed again
+        # print("serve brake turned off")
         # called from train controller container when train sends new values
         # perform all calculations for new values here and update output array
 
@@ -438,18 +489,22 @@ class TrainController:
         if num == 1:
             print("train controller sw.py: update type 1 values (authority safe speed): authority and cmd speed")
             self.vitalAuth = inputs[0]  # authority is distance to destination
-            print(self.vitalAuth)
-            print(inputs[0])
+            # print(self.vitalAuth)
+            # print(inputs[0])
+            print(inputs)
+            print()
 
             self.cmdSpeed = inputs[1]
-            print(self.cmdSpeed)
-            self.speedcheck()
+            # print(self.cmdSpeed)
+            # self.speedcheck()
 
             # distance control
             self.vitalitycheck()
-            self.authority()
+            # self.authority()
         elif num == 2:
             print("train controller sw.py: update type 2 values (track info): polarity, underground, beacon")
+            print(inputs)
+            print()
             # check if in new block and update block values
             self.polaritycontrol(inputs[0])
             self.isUnderground = inputs[1]
@@ -457,28 +512,62 @@ class TrainController:
             if self.isUnderground:
                 self.intlightson()
                 self.extlightson()
-            elif not self.isUnderground:
+            elif not self.isUnderground and not self.extLightsByDriver and not self.intLightsByDriver:
                 self.intlightsoff()
                 self.extlightsoff()
+
+            if self.intLights:
+                print("train controller sw.py: cabin lights on")
+            elif not self.intLights:
+                print("train controller sw.py: cabin lights off")
+
+            if self.extLights:
+                print("train controller sw.py: headlights on")
+            elif not self.extLights:
+                print("train controller sw.py: headlights off")
+            print()
 
             self.beacon = inputs[2]
             self.parsebeacon(self.beacon)
         elif num == 3:
             print("train controller sw.py: update type 3 values (train info): actual speed, passenger e-brake")
+            print(inputs)
+
             self.actualSpeed = inputs[0]
             self.passEBrake = inputs[1]
-            self.ebrakecontrol(self.passEBrake)
+            print
+            #self.ebrakecontrol(self.passEBrake)
+
+            if self.passEBrake:
+                self.eBrake = True
 
             # power stuff
             self.powercontrol()
+            print(f'train controller sw.py: power is {self.power}')
 
-        print(f'train controller sw.py: power is {self.power}')
+            # if self.atStation and self.actualSpeed == 0 and self.power == 0:
+            #     self.doorcontrol()
+            #     print(f'entered station: doorstate is {self.doorState}')
+            #     SystemTime.pause()
+            #     time.sleep(60/SystemTime.scale)
+            #     self.doorcontrol()
+            #     print(f'leaving station: doorstate is: {self.doorState}')
 
         return
 
+    def arriveatstation(self, is_at_station):
+        if is_at_station and self.actualSpeed == 0 and self.power == 0:
+            print(f'entered station at: {SystemTime.time()}')
+            self.doorcontrol()
+            print(f'opening doors, door state is {self.doorState}')
+            # time.sleep(6)  # waits at station for 60 seconds, set to 6 seconds for individual module purposes
+            self.doorcontrol()
+            print(f'closing doors, door state is: {self.doorState}')
+            print(f'left station at: {SystemTime.time()}')
+
     def old_updater(self, inputs):
         print("train controller sw.py: train controller values being updated in old")
-        self.servBrake = 0  # turn off service brake, will be turned on if needed again
+        self.servBrake = False  # turn off service brake, will be turned on if needed again
         # print("serve brake turned off")
         # called from train controller container when train sends new values
         # perform all calculations for new values here and update output array
@@ -533,6 +622,6 @@ class TrainController:
         # call update function to inject new values to train model
         # 8 el list + cabin temp
 
-        outputs = [self.cmdSpeed, self.power, self.servBrake, self.eBrake, self.doorSide, self.makeAnnouncement,
+        outputs = [self.cmdSpeed, self.power, self.servBrake, self.eBrake, self.doorSide, self.station,
                    self.intLights, self.extLights, self.cabinTemp]
         return outputs
